+++ DemoApp/Shared/StreamChatWrapper.swift
@@ -147,20 +147,7 @@ extension StreamChatWrapper {
             return
         }
 
-        let currentUserController = client.currentUserController()
-        currentUserController.synchronize()
-        if let deviceId = currentUserController.currentUser?.currentDevice?.id {
-            currentUserController.removeDevice(id: deviceId) { error in
-                if let error = error {
-                    log.error("Removing the device failed with an error \(error)")
-                }
-
-                client.logout(completion: completion)
-            }
-        } else {
-            log.error("No deviceId has been found from the current user.")
-            client.logout(completion: completion)
-        }
+        client.logout(completion: completion)
     }
 }