+++ Sources/StreamChat/Database/DTOs/ChannelDTO.swift
@@ -7,6 +7,9 @@ import Foundation
 
 @objc(ChannelDTO)
 class ChannelDTO: NSManagedObject {
+    // The cid without the channel type.
+    @NSManaged var id: String?
+    // The channel id which includes channelType:channelId.
     @NSManaged var cid: String
     @NSManaged var name: String?
     @NSManaged var imageURL: URL?
@@ -62,7 +65,11 @@ class ChannelDTO: NSManagedObject {
     @NSManaged var messages: Set<MessageDTO>
     @NSManaged var pinnedMessages: Set<MessageDTO>
     @NSManaged var reads: Set<ChannelReadDTO>
+
+    /// Helper properties used for sorting channels with unread counts of the current user.
     @NSManaged var currentUserUnreadMessagesCount: Int32
+    @NSManaged var hasUnreadSorting: Int16
+
     @NSManaged var watchers: Set<UserDTO>
     @NSManaged var memberListQueries: Set<ChannelMemberListQueryDTO>
     @NSManaged var previewMessage: MessageDTO?
@@ -79,12 +86,13 @@ class ChannelDTO: NSManagedObject {
         }
 
         // Update the unreadMessagesCount for the current user.
-        // At the moment this computed property is used for `hasUnread` automatic channel list filtering.
+        // At the moment this computed property is used for `hasUnread` and `unreadCount` automatic channel list filtering.
         if let currentUserId = managedObjectContext?.currentUser?.user.id {
             let currentUserUnread = reads.first(where: { $0.user.id == currentUserId })
             let newUnreadCount = currentUserUnread?.unreadMessageCount ?? 0
             if newUnreadCount != currentUserUnreadMessagesCount {
                 currentUserUnreadMessagesCount = newUnreadCount
+                hasUnreadSorting = newUnreadCount > 0 ? 1 : 0
             }
         }
 
@@ -105,7 +113,7 @@ class ChannelDTO: NSManagedObject {
                 newestMessageAt = nil
             }
         }
-
+        
         // Update the date for sorting every time new message in this channel arrive.
         // This will ensure that the channel list is updated/sorted when new message arrives.
         // Note: If a channel is truncated, the server will update the lastMessageAt to a minimum value, and not remove it.
@@ -240,6 +248,7 @@ extension NSManagedObjectContext {
             dto.extraData = Data()
         }
         dto.typeRawValue = payload.typeRawValue
+        dto.id = payload.cid.id
         dto.config = payload.config.asDTO(context: self, cid: dto.cid)
         if let ownCapabilities = payload.ownCapabilities {
             dto.ownCapabilities = ownCapabilities