+++ Sources/StreamChat/WebSocketClient/Events/MessageEvents.swift
@@ -310,6 +310,7 @@ private extension UserPayload {
             isBanned: isBanned,
             isFlaggedByCurrentUser: false,
             userRole: role,
+            teamsRole: teamsRole,
             createdAt: createdAt,
             updatedAt: updatedAt,
             deactivatedAt: deactivatedAt,